+++ src/ioc2rpz_rest.erl
@@ -257,9 +257,9 @@ srv_mgmt(Req, State, Format) when State#state.op == update_rpz -> %Update an RPZ
 		[] -> false
 	end,
 	{Body,Req0} = case {ZoneS, Format} of
-		{true,json} -> {io_lib:format("{\"status\":\"ok\",\"msg\":\"RPZ ~s will be updated\"}\n",[RPZ]),Req};
+		{true,json} -> {io_lib:format("{\"status\":\"ok\",\"msg\":\"RPZ ~s will be updated\"}\n",[ioc2rpz_fun:json_escape(RPZ)]),Req};
 		{true,txt} -> {io_lib:format("status: ok\nmsg: RPZ ~s will be updated\n",[RPZ]),Req};
-		{false,json} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(136),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("{\"status\":\"error\",\"msg\":\"RPZ ~s not found\"}\n",[RPZ]),cowboy_req:reply(520, Req)};
+		{false,json} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(136),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("{\"status\":\"error\",\"msg\":\"RPZ ~s not found\"}\n",[ioc2rpz_fun:json_escape(RPZ)]),cowboy_req:reply(520, Req)};
 		{false,txt} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(136),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("status: error\nmsg: RPZ ~s not found\n",[RPZ]),cowboy_req:reply(520, Req)}
 	end,
 	{Body, Req0, State};
@@ -317,9 +317,9 @@ srv_mgmt(Req, State, Format) when State#state.op == get_rpz -> % Get RPZ
   #{type := Type} = cowboy_req:match_qs([{type, [], <<"both">>}], Req),
 %  erlang:display(Type),
 	{Body,Req0} = case {Data, Format, Type} of
-		{[],json,_} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(138),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("{\"status\":\"error\",\"msg\":\"RPZ ~s not found\"}\n",[RPZ]),cowboy_req:reply(520, Req)};
+		{[],json,_} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(138),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("{\"status\":\"error\",\"msg\":\"RPZ ~s not found\"}\n",[ioc2rpz_fun:json_escape(RPZ)]),cowboy_req:reply(520, Req)};
 		{[],txt,_} -> ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(138),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]), {io_lib:format("status: error\nmsg: RPZ ~s not found\n",[RPZ]),cowboy_req:reply(520, Req)};
-		{_,json,_} -> {io_lib:format("{\"status\":\"ok\",\"rpz\":\"~s\",\"iocs\":[~s]}\n",[RPZ,ioc2jsonarr(Data,binary_to_list(Type))]),Req};
+		{_,json,_} -> {io_lib:format("{\"status\":\"ok\",\"rpz\":\"~s\",\"iocs\":[~s]}\n",[ioc2rpz_fun:json_escape(RPZ),ioc2jsonarr(Data,binary_to_list(Type))]),Req};
 %		{_,txt} -> {lists:flatten([ io_lib:format("~s,~s\n",[binary_to_list(X),Type]) || [X,_Ser,_Exp,Type] <- Data]),Req}
     {_,txt,<<"fqdn">>} -> {lists:flatten([ io_lib:format("~s\n",[binary_to_list(X)]) || [X,_Ser,_Exp,"fqdn"] <- Data]),Req};
     {_,txt,<<"ip">>} -> {lists:flatten([ io_lib:format("~s\n",[binary_to_list(X)]) || [X,_Ser,_Exp,"ip"] <- Data]),Req};
@@ -504,7 +504,7 @@ tuple_to_json({Name,Value}) when is_integer(Value)->
     io_lib:format("{\"~s\":~b}",[Name,Value]);
 
 tuple_to_json({Name,Value}) ->
-    io_lib:format("{\"~s\":\"~s\"}",[Name,Value]);
+    io_lib:format("{\"~s\":\"~s\"}",[Name,ioc2rpz_fun:json_escape(Value)]);
 
 tuple_to_json(REST) ->
 	Res=mtuple_to_json([],REST),
@@ -514,13 +514,13 @@ mtuple_to_json([],[{Name,Value}|REST]) when is_integer(Value)->
     mtuple_to_json(io_lib:format("\"~s\":~b",[Name,Value]),REST);
 
 mtuple_to_json([],[{Name,Value}|REST]) ->
-    mtuple_to_json(io_lib:format("\"~s\":\"~s\"",[Name,Value]),REST);
+    mtuple_to_json(io_lib:format("\"~s\":\"~s\"",[Name,ioc2rpz_fun:json_escape(Value)]),REST);
 
 mtuple_to_json(Val,[{Name,Value}|REST]) when is_integer(Value)->
     mtuple_to_json(Val++io_lib:format(",\"~s\":~b",[Name,Value]),REST);
 
 mtuple_to_json(Val,[{Name,Value}|REST]) ->
-    mtuple_to_json(Val++io_lib:format(",\"~s\":\"~s\"",[Name,Value]),REST);
+    mtuple_to_json(Val++io_lib:format(",\"~s\":\"~s\"",[Name,ioc2rpz_fun:json_escape(Value)]),REST);
 
 mtuple_to_json(Val,[]) ->
     Val.
@@ -532,13 +532,13 @@ ioc2jsonarr(IOCs,Type) ->
 
 ioc2jsonarr([],[[IOC,_,_,IType]|REST],Type) ->
   if (IType == Type) or (Type == "both") ->
-    ioc2jsonarr(io_lib:format("\"~s\"",[binary_to_list(IOC)]),REST,Type);
+    ioc2jsonarr(io_lib:format("\"~s\"",[ioc2rpz_fun:json_escape(IOC)]),REST,Type);
     true -> ioc2jsonarr([],REST,Type)
   end;
 
 ioc2jsonarr(Resp,[[IOC,_,_,IType]|REST],Type) ->
   if (IType == Type) or (Type == "both") ->
-    ioc2jsonarr(io_lib:format("\"~s\",",[binary_to_list(IOC)])++Resp,REST,Type);
+    ioc2jsonarr(io_lib:format("\"~s\",",[ioc2rpz_fun:json_escape(IOC)])++Resp,REST,Type);
     true -> ioc2jsonarr(Resp,REST,Type)
   end;